Split classes into seperate files
