+++ tests/X.PagedList.EF.Tests/Blog.cs
@@ -1,0 +1,9 @@
+﻿namespace X.PagedList.EF.Tests
+{
+    class Blog
+    {
+        public int BlogId { get; set; }
+        public string Name { get; set; }
+        public string Url { get; set;  }
+    }
+}
